Implemented GlobalPathway
