+++ UnityProject2018/Assets/Scenes/devNetwork-Main.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 37e76ce3f87a97c45b658f2de9c6b7c1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
