+++ UnityProject2018/Assets/Scripts/QueryServices/PrefabService.cs
@@ -41,7 +41,7 @@ public class PrefabService : MonoBehaviour
         if (obj != null) {
             if(obj.GetComponentInChildren<NodeDataDisplay>().nodeData == null) {
                 obj.GetComponentInChildren<NodeDataDisplay>().nodeData = node;
-                Debug.Log("Attached node " + nodeName);
+                // Debug.Log("Attached node " + nodeName);
             }
         } else {
             Debug.LogError("Node scriptable object not connected to prefab: " + nodeName);
