+++ UnityProject2018/Assets/Scripts/QueryServices/QueryService.cs
@@ -12,7 +12,7 @@ public class QueryService : MonoBehaviour
 static Dictionary<string,EdgeSO> EdgeSOs = new Dictionary<string, EdgeSO>();
 static Dictionary<string,NodeSO> NodeSOs = new Dictionary<string, NodeSO>();
 public static Dictionary<string,PathwaySO> PathwaySOs = new Dictionary<string, PathwaySO>();
-static PathwaySO GlobalPathway;
+static ConnectionsSO globalPathway;
 
 public TextAsset queryxml;
 
@@ -90,7 +90,7 @@ void SerializeAndCreate()
     
     WikibaseResult result = JsonUtility.FromJson<WikibaseResult>(xmlToString);
 
-    GlobalPathway = ScriptableObject.CreateInstance<PathwaySO>();
+    globalPathway = ScriptableObject.CreateInstance<ConnectionsSO>();
     foreach( WikibaseBinding item in result.results.bindings){
                 NodeSOInit(item);
             }
@@ -100,12 +100,14 @@ void SerializeAndCreate()
 
 // fill active pathways in statuscontroller after querying 
 void PathwaysToActive(){
+    FillPathwayList(); //TODO: list usage seems to be unused, added for testing MockSearch()
     GameObject.Find("StatusController").GetComponent<StatusController>().activePathways.Clear();
     foreach(KeyValuePair<string,PathwaySO> pair in PathwaySOs)
         {
             GameObject.Find("StatusController").GetComponent<StatusController>().activePathways.Add(pair.Value);
 
         }
+    GameObject.Find("StatusController").GetComponent<StatusController>().globalPathway = globalPathway;
 }
 
 
@@ -184,7 +186,9 @@ public void NodeSOInit(WikibaseBinding item){
     currentPathway.AddNode(currentNode);
     currentPathway.AddEdge(currentNode,currentEdge);
 
-    // 
+    // add nodes and edges to global pathway
+    globalPathway.AddNode(currentNode);
+    globalPathway.AddEdge(currentNode, currentEdge);
     
 }
