+++ UnityProject2018/Assets/Scripts/ScriptableObjects/DataTemplates/ConnectionsSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 51e249ae233a6b24484c098488622d45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
